# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FleetSaaS school bus tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /passengers:
    get:
      operationId: listPassengers
      tags: [passengers]
      summary: List all passengers for today's trip checklist
      parameters:
        - in: query
          name: phone
          schema:
            type: string
          description: Filter passengers to those whose parent phone matches this value (used for parent portal scoping)
      responses:
        "200":
          description: Passenger list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Passenger"
    post:
      operationId: createPassenger
      tags: [passengers]
      summary: Register a new passenger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PassengerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passenger"
  /passengers/{id}:
    delete:
      operationId: deletePassenger
      tags: [passengers]
      summary: Remove a passenger
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
    get:
      operationId: getPassenger
      tags: [passengers]
      summary: Get passenger by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Passenger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passenger"
    patch:
      operationId: updatePassenger
      tags: [passengers]
      summary: Update passenger profile or photo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PassengerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passenger"
  /passengers/{id}/send-boarding-otp:
    post:
      operationId: sendBoardingOtp
      tags: [passengers]
      summary: Generate a server-side boarding OTP for a passenger (driver action)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OTP generated (sent to parent via SMS in production; returned as demoCode in simulation)
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
                  demoCode:
                    type: string
                    description: Demo only — the generated OTP for testing (not shown in production UI)
  /passengers/{id}/qr-token:
    get:
      operationId: getPassengerQrToken
      tags: [passengers]
      summary: Get a signed QR token for a passenger (parent action) — encodes passenger ID with HMAC signature
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Signed QR token payload (display as QR code in parent app)
          content:
            application/json:
              schema:
                type: object
                required: [token]
                properties:
                  token:
                    type: string
                    description: Signed token string to embed in a QR code
  /passengers/{id}/board:
    post:
      operationId: boardPassenger
      tags: [passengers]
      summary: Mark passenger as boarded (driver action) — requires valid OTP or QR token
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                otp:
                  type: string
                  description: The OTP entered by the driver, validated server-side
                qrToken:
                  type: string
                  description: Signed QR token scanned from the parent's device (alternative to OTP)
      responses:
        "200":
          description: Updated passenger with boarding status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passenger"
        "401":
          description: Invalid or expired OTP / QR token
  /passengers/{id}/unboard:
    post:
      operationId: unboardPassenger
      tags: [passengers]
      summary: Mark passenger as unboarded (driver action)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated passenger with pending status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passenger"
  /passengers/{id}/leave:
    post:
      operationId: markPassengerLeave
      tags: [passengers]
      summary: Mark passenger as on leave (parent or admin action)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated passenger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passenger"

  # ─── TRIPS ──────────────────────────────────────────────────────────────────
...
  /routes:
    get:
      operationId: listRoutes
      tags: [routes]
      summary: List all routes for this tenant
      responses:
        "200":
          description: Route list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Route"
    post:
      operationId: createRoute
      tags: [routes]
      summary: Create a new route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteInput"
      responses:
        "201":
          description: Created route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
  /routes/{id}:
    patch:
      operationId: updateRoute
      tags: [routes]
      summary: Update route name or assigned driver/vehicle (fleet swap)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteUpdate"
      responses:
        "200":
          description: Updated route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
    delete:
      operationId: deleteRoute
      tags: [routes]
      summary: Delete a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    RouteInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        driverId:
          type: integer
        vehicleId:
          type: integer
        departureTime:
          type: string
        avgSpeedKmh:
          type: integer
...
    Passenger:
      type: object
      required: [id, name, role, status, stationId]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, staff, driver]
        status:
          type: string
          enum: [pending, boarded, absent, leave]
        stationId:
          type: integer
        routeId:
          type: ["integer", "null"]
        stationName:
          type: ["string", "null"]
        boardedAt:
          type: ["string", "null"]
          format: date-time
        liveToday:
          type: integer
          description: 1 if the passenger confirmed they are riding today, 0 otherwise
        quickMessage:
          type: ["string", "null"]
          description: Last quick status message sent by the passenger
        className:
          type: ["string", "null"]
          description: "Class/grade the student is enrolled in (e.g. Class 10, Grade 8)"
        section:
          type: ["string", "null"]
          description: "Section within the class (e.g. A, B, Science)"
        rollNumber:
          type: ["string", "null"]
          description: School roll / admission number
        faculty:
          type: ["string", "null"]
          description: "Faculty or stream, if applicable (e.g. Science, Management, Humanities)"
        designation:
          type: ["string", "null"]
          description: "Staff designation (e.g. Teacher, Accountant, School Staff, or a custom value)"
        parentName:
          type: ["string", "null"]
          description: Parent or guardian full name
        gender:
          type: ["string", "null"]
          description: "Passenger gender (Male / Female / Other)"
    PassengerInput:
      type: object
      required: [name, role, stationId]
      properties:
        name:
          type: string
        phone:
          type: string
          description: Contact / parent phone number (used for OTP login)
        photoUrl:
          type: string
        parentName:
          type: string
          description: Parent or guardian full name
        gender:
          type: string
          description: "Passenger gender (Male / Female / Other)"
        role:
          type: string
          enum: [student, staff, driver]
        stationId:
          type: integer
        routeId:
          type: integer
        className:
          type: string
          description: "Class/grade the student is enrolled in (e.g. Class 10, Grade 8)"
        section:
          type: string
          description: "Section within the class (e.g. A, B, Science)"
        rollNumber:
          type: string
          description: School roll / admission number
        faculty:
          type: string
          description: "Faculty or stream, if applicable (e.g. Science, Management, Humanities)"
        designation:
          type: string
          description: "Staff designation (e.g. Teacher, Accountant, School Staff, or a custom value)"
    PassengerUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        photoUrl:
          type: string
        stationId:
          type: integer
        routeId:
          type: ["integer", "null"]
        liveToday:
          type: integer
        quickMessage:
          type: string
        className:
          type: string
        section:
          type: string
        rollNumber:
          type: string
        faculty:
          type: string
        designation:
          type: string
        parentName:
          type: string
        gender:
          type: string
...